+++ lib/src/interfaces/handler.dart
@@ -1,0 +1,4 @@
+/// handler for handlers annotiations
+class HandlerInfo {
+  const HandlerInfo();
+}